Edit Clients

+++ BHelp/Controllers/ClientsController.cs
@@ -1,10 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Data;
 using System.Data.Entity;
 using System.Linq;
 using System.Net;
-using System.Web;
 using System.Web.Mvc;
 using BHelp.DataAccessLayer;
 using BHelp.Models;
@@ -19,7 +17,43 @@ namespace BHelp.Controllers
         // GET: Clients
         public ActionResult Index()
         {
-            return View(db.Clients.ToList());
+            var clientList = db.Clients.Where(c => c.Active).OrderBy(c => c.LastName).ToList();
+            var clientView = new List<ClientViewModel>();
+            foreach (var client in clientList)
+            {
+                var household = new ClientViewModel()
+                {
+                    Id = client.Id,
+                    FirstName = client.FirstName,
+                    LastName = client.LastName,
+                    Age = AppRoutines.GetAge(client.DateOfBirth,DateTime.Today),
+                    StreetNumber = client.StreetNumber,
+                    StreetName = client.StreetName,
+                    StreetToolTip = client.StreetName.Replace(" ", "\u00a0"),
+                    City = client.City,
+                    CityToolTip = client.City.Replace(" ", "\u00a0"),
+                    Zip = client.Zip,
+                    Phone = client.Phone,
+                    PhoneToolTip = client.Phone.Replace(" ", "\u00a0"),
+                    Notes = client.Notes,
+                    // (full length on mouseover)    \u00a0 is the Unicode character for NO-BREAK-SPACE.
+                    NotesToolTip = client.Notes.Replace(" ", "\u00a0")
+                };
+                var s = household.StreetName; // For display, abbreviate to 10 characters:           
+                s = s.Length <= 10 ? s : s.Substring(0, 10) + "...";
+                household.StreetName = s;
+                s = household.City; // For display, abbreviate to 11 characters:           
+                s = s.Length <= 11 ? s : s.Substring(0, 11) + "...";
+                household.City = s;
+                s = household.Phone; // For display, abbreviate to 12 characters:           
+                s = s.Length <= 12 ? s : s.Substring(0, 12) + "...";
+                household.Phone = s;
+                s = household.Notes; // For display, abbreviate to 12 characters:           
+                s = s.Length <= 12 ? s : s.Substring(0, 12) + "...";
+                household.Notes = s;
+                clientView.Add(household);
+            }
+            return View(clientView);
         }
 
         // GET: Clients/Details/5
@@ -67,6 +101,8 @@ namespace BHelp.Controllers
                     LastName = client.LastName,
                     StreetNumber = client.StreetNumber,
                     StreetName = client.StreetName,
+                    Phone = client.Phone,
+                    City = client.City,
                     Zip = client.Zip,
                     Notes = client.Notes
                 };
@@ -118,10 +154,19 @@ namespace BHelp.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "Id,Active,FirstName,LastName,DateOfBirth,StreetNumber,StreetName,City,Zip,Phone,Notes")] Client client)
+        public ActionResult Edit([Bind(Include = "Id,Active,FirstName,LastName,DateOfBirth," +
+                           "StreetNumber,StreetName,City,Zip,Phone,Notes")] Client client)
         {
             if (ModelState.IsValid)
             {
+                if (client.FirstName == null) { client.FirstName = ""; }
+                if (client.LastName == null) { client.LastName = ""; }
+                if (client.StreetNumber == null) { client.StreetNumber = ""; }
+                if (client.StreetName == null) { client.StreetName = ""; }
+                if (client.City == null) { client.City = ""; }
+                if (client.Phone == null) { client.Phone = ""; }
+                if (client.Notes == null) { client.Notes = ""; }
+                client.DateOfBirth=DateTime.Today.AddYears(-client.Age);
                 db.Entry(client).State = EntityState.Modified;
                 db.SaveChanges();
                 return RedirectToAction("Index");
@@ -150,7 +195,7 @@ namespace BHelp.Controllers
         public ActionResult DeleteConfirmed(int id)
         {
             Client client = db.Clients.Find(id);
-            db.Clients.Remove(client);
+            if (client != null) db.Clients.Remove(client);
             db.SaveChanges();
             return RedirectToAction("Index");
         }
